refactor(frontend): constants closer to sdk's
