+++ packages/frontend/constants/asset.ts
@@ -2,7 +2,3 @@ export const DEFAULT_LTV_MAX = 75;
 export const DEFAULT_LTV_TRESHOLD = 100;
 export const LTV_RECOMMENDED_DECREASE = 20;
 export const MINIMUM_DEBT_AMOUNT = 1;
-
-export const DEFAULT_SLIPPAGE = 30;
-
-export const BALANCE_POLLING_INTERVAL = 15000;
