+++ packages/frontend/constants/common.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_SLIPPAGE = 30;
+
+export const BALANCE_POLLING_INTERVAL = 15000;
